renamed show_format.tsx, corresponding folder and component
