+++ components/advanced_text_editor/show_formatting/index.ts
@@ -1,4 +1,4 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-export {default} from './show_format';
+export {default} from './show_formatting';
